Chenges - buggy, but need help from Josh Pepper
Something is wrong with the reactive variables - lines 16 and 17. How it runs frametest is weird. When I try to use the 'CrossFilter" input slot from the UI, the fourth filter on the map (Show where this country works) here is an example of the error I currently see: 

Warning: Unhandled error in observer: attempt to select less than one element
observe({
    pal <- colorpal()
    colorBy <- input$MapFilter
    leafletProxy("map", data = filteredData()) %>% clearShapes() %>% 
        clearControls() %>% addPolygons(stroke = NULL, smoothFactor = 0.5, 
        color = ~pal(filteredData()@data[, colorBy]), opacity = 0.9, 
        popup = ~paste("<strong>", colorBy, ":</strong>", filteredData()@data[, 
            colorBy], "<strong>Country:</strong>", NAME)) %>% 
        addLegend(title = colorBy, pal = colorpal(), values = filteredData()@data[, 
            colorBy], position = "bottomright")
})
Error in .subset2(x, i, exact = exact) : 
  attempt to select less than one element

Once this part of the cod works, it will be much easier to code the rest of the materials. I feel it's a simple mistake that I don't know how to fix.