DataFilter works, need to clean code and work out allocating bugs
So, apparently the filtering code works but under certain circumstances.
I want to make sure it works in all instances, and it works as
effectively as manually filtering ourselves.

Also, Datafilter (which can now start to be the replacement for
frametest) is not implemented in the Shiny code. I'll explain further in
a further email/commit.